Removed legacy- and unused code in neuralnet/layer
Several fixes on old Header-Guards, variable declarations.
Removed all methods that are still unused, for they are not necessary for the program and thus need not be tested.

+++ src/neuralnet/layers/functionlayers/ActivationLayer.cpp
@@ -48,7 +48,7 @@ void ActivationLayer::setPlatform(Platform *platform) {
     this->functionSet = true;
 }
 
-long long int ActivationLayer::getDifficulty() {
+int ActivationLayer::getDifficulty() {
     inputWrapper = new DataWrapper(inputDimensions);
     if (this->difficulty == 0)
         this->difficulty = this->inputWrapper->getNumElements(); // Linear on input
