+++ src/neuralnet/layers/functionlayers/ActivationLayer.h
@@ -47,7 +47,7 @@ public:
 
     void setPlatform(Platform *platform) override;
 
-    long long int getDifficulty() override;
+    int getDifficulty() override;
 
 };
