+++ src/neuralnet/layers/functionlayers/LossLayer.h
@@ -39,6 +39,6 @@ public:
 
     void setPlatform(Platform *platform) override;
 
-    long long int getDifficulty() override;
+    int getDifficulty() override;
 };
