+++ src/neuralnet/layers/naive/NaiveLayer.cpp
@@ -31,7 +31,7 @@ void NaiveLayer::setPlatform(Platform *platform) {
     this->functionSet = true;
 }
 
-long long int NaiveLayer::getDifficulty() {
+int NaiveLayer::getDifficulty() {
     return 1; // Naive layers don't perfom computation
 }
